Move the assertion helpers into a utils folder and use in other tests
